requirements.txt
lxml==3.4.4
passlib==1.6.5
python-docx==0.8.5
selenium==2.48.0
wheel==0.24.0
pythonwx
